Fixed RAIN-1057 and added data-localization In Label and divided Inbox by components
