+++ web/rainmaker/dev-packages/egov-ui-kit-dev/src/components/Label/index.js
@@ -2,14 +2,14 @@ import React from "react";
 import PropTypes from "prop-types";
 import "./index.css";
 
-const labelText = (label, labelStyle, labelClassName, required,secondaryText) => {
+const labelText = (label, labelStyle, labelClassName, required, secondaryText) => {
   return label && label.length ? (
-    <div className={`label-text ${labelClassName}`} style={labelStyle}>
-      {label} {secondaryText} 
-      {required && <span style={{color: "red"}}> *</span>}
+    <div data-localization={label} className={`label-text ${labelClassName}`} style={labelStyle}>
+      {label} {secondaryText}
+      {required && <span style={{ color: "red" }}> *</span>}
     </div>
   ) : (
-    "" 
+    ""
   );
 };
 
@@ -27,7 +27,7 @@ const Label = ({
   buttonLabel = false,
   id,
   required,
-  secondaryText=""
+  secondaryText = "",
 }) => {
   let additionalStyles = {};
 
@@ -56,7 +56,7 @@ const Label = ({
 
   return (
     <div id={id} style={containerStyle} className={buttonLabel ? `button-label-container ${className}` : `label-container ${className}`}>
-      {labelText(label, labelStyle, labelClassName, required,secondaryText)}
+      {labelText(label, labelStyle, labelClassName, required, secondaryText)}
     </div>
   );
 };
