+++ web/rainmaker/packages/employee/src/modules/employee/Inbox/components/Table/index.scss
